backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Smart Travel Planner backend — FastAPI + LangGraph agent + pgvector RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14",
    "fastapi>=0.136.1",
    "httpx>=0.27",
    "joblib>=1.5.3",
    "langchain>=1.2.15",
    "langgraph>=1.1.10",
    "pandas>=3.0.2",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.2",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.6",
    "scikit-learn>=1.8.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.46.0",
    "voyageai>=0.3.4",
    "langchain-google-genai>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "pydantic[email]>=2.13.3",
    "email-validator>=2.0",
    "bcrypt==4.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.9",
    "seaborn>=0.13.2",
]

[tool.alembic]
script_location = "alembic"
